requirement.txt
absl-py==1.3.0
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.6.1
mediapipe==0.8.11
numpy==1.23.4
opencv-contrib-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.3
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
